extract tokens in a separate file
